requirements.txt
requests>=1.0.4
tqdm>=3.8.0
futures==2.2
tornado>=4.1
pyyaml>=3.11
pycurl